--- a/components/form/Internal/Validate/ValidationAttribute/NumberAttribute.cs
+++ b/components/form/Internal/Validate/ValidationAttribute/NumberAttribute.cs
@@ -2,35 +2,18 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.ComponentModel.DataAnnotations;
-using System.Globalization;
-
 namespace AntDesign.Internal.Form.Validate;
 
-internal sealed class NumberAttribute(decimal number) : ValidationAttribute
+internal sealed class NumberAttribute(decimal number) : NumberCompareAttribute(number, 0)
 {
-    internal decimal Number { get; } = number;
+}
 
-    public override string FormatErrorMessage(string name)
-    {
-        return string.Format(CultureInfo.CurrentCulture, ErrorMessageString, name, Number);
-    }
+internal sealed class NumberMinAttribute(decimal min) : NumberCompareAttribute(min, 1)
+{
+    // internal decimal Min => Number;
+}
 
-    public override bool IsValid(object value)
-    {
-        if (value == null)
-        {
-            return true;
-        }
-
-        return value switch
-        {
-            decimal valueAsDecimal => Number == valueAsDecimal,
-            double valueAsDouble => Number == (decimal)valueAsDouble,
-            float valueAsFloat => Number == (decimal)valueAsFloat,
-            int valueAsInt => Number == (decimal)valueAsInt,
-            long valueAsLong => Number == (decimal)valueAsLong,
-            _ => false,
-        };
-    }
+internal sealed class NumberMaxAttribute(decimal max) : NumberCompareAttribute(max, -1)
+{
+    // internal decimal Max => Number;
 }