<!--
First of all, thank you for your contribution! 😄

New feature please send pull request to feature branch, and rest to master branch.
Pull request will be merged after one of collaborators approve.
Please makes sure that these form are filled before submitting your pull request, thank you!

[[中文版模板 / Chinese template](https://github.com/ant-design-blazor/ant-design-blazor/blob/master/.github/PULL_REQUEST_TEMPLATE/pr_cn.md)]

-->

### 🤔 This is a ...

- [ ] New feature
- [ ] Bug fix
- [ ] Site / documentation update
- [ ] Demo update
- [ ] Component style update
- [ ] Bundle size optimization
- [x] Performance optimization
- [x] Refactoring
- [ ] Code style optimization
- [ ] Test Case
- [ ] Branch merge
- [ ] Other (about what?)

### 🔗 Related issue link

<!--
1. Describe the source of requirement, like related issue link.
-->

### 💡 Background and solution

<!--
1. Describe the problem and the scenario.
2. GIF or snapshot should be provided if includes UI/interactive modification.
3. How to fix the problem, and list final API implementation and usage sample if that is an new feature.
-->

This PR tries to optimize the internal number validation attributes used by Form component.

Here is a micro-benchmark,

```

BenchmarkDotNet v0.15.1, Windows 11 (10.0.26100.4351/24H2/2024Update/HudsonValley)
Intel Core i7-8700 CPU 3.20GHz (Max: 3.19GHz) (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET SDK 9.0.301
  [Host]     : .NET 9.0.6 (9.0.625.26613), X64 RyuJIT AVX2
  DefaultJob : .NET 9.0.6 (9.0.625.26613), X64 RyuJIT AVX2


```
| Method   | value     | Mean      | Error    | StdDev   | Ratio | Gen0   | Allocated | Alloc Ratio |
|--------- |---------- |----------:|---------:|---------:|------:|-------:|----------:|------------:|
| **OldValid** | **999**       | **303.01 ns** | **3.205 ns** | **2.677 ns** |  **1.00** | **0.0749** |     **472 B** |        **1.00** |
| NewValid | 999       |  30.17 ns | 0.629 ns | 0.772 ns |  0.10 | 0.0127 |      80 B |        0.17 |
|          |           |           |          |          |       |        |           |             |
| **OldValid** | **999.99**    | **530.97 ns** | **2.137 ns** | **1.894 ns** |  **1.00** | **0.0763** |     **480 B** |        **1.00** |
| NewValid | 999.99    |  41.59 ns | 0.206 ns | 0.192 ns |  0.08 | 0.0127 |      80 B |        0.17 |
|          |           |           |          |          |       |        |           |             |
| **OldValid** | **999.999**   | **624.46 ns** | **3.399 ns** | **3.180 ns** |  **1.00** | **0.0782** |     **496 B** |        **1.00** |
| NewValid | 999.999   |  44.91 ns | 0.817 ns | 0.682 ns |  0.07 | 0.0127 |      80 B |        0.16 |
|          |           |           |          |          |       |        |           |             |
| **OldValid** | **999.99999** | **618.39 ns** | **3.373 ns** | **3.155 ns** |  **1.00** | **0.0839** |     **528 B** |        **1.00** |
| NewValid | 999.99999 |  29.06 ns | 0.266 ns | 0.248 ns |  0.05 | 0.0127 |      80 B |        0.15 |

The source codes:

```csharp
[MemoryDiagnoser]
public class ValidationSample
{
    private const decimal Number = 1000;

    [Benchmark(Baseline = true)]
    [ArgumentsSource(nameof(Data))]
    public bool OldValid(object value) => new OldNumberMaxAttribute(Number).IsValid(value);

    [Benchmark]
    [ArgumentsSource(nameof(Data))]
    public bool NewValid(object value) => new NumberMaxAttribute(Number).IsValid(value);

    public IEnumerable<object> Data()
    {
        yield return 999.999d;
        yield return 999.99f;
        yield return 999;
        yield return 999.99999m;
    }

    internal abstract class NumberCompareAttribute(decimal number, int comparison) : ValidationAttribute
    {
        private static readonly Dictionary<Type, Func<object, decimal>> _defaultDecimalConverters = new()
    {
        { typeof(decimal), static value => (decimal)value },
        { typeof(double), static value => (decimal)(double)value },
        { typeof(float), static value => (decimal)(float)value },
        { typeof(int), static value => (int)value },
        { typeof(long), static value => (long)value },
        { typeof(short), static value => (short)value },
        { typeof(sbyte), static value => (sbyte)value },
        { typeof(ulong), static value => (ulong)value },
        { typeof(uint), static value => (uint)value },
        { typeof(ushort), static value => (ushort)value },
        { typeof(byte), static value => (byte)value },
        { typeof(decimal?), static value => ((decimal?)value).Value },
        { typeof(double?), static value => (decimal)((double?)value).Value },
        { typeof(float?), static value => (decimal)((float?)value).Value },
        { typeof(int?), static value => ((int?)value).Value },
        { typeof(long?), static value => ((long?)value).Value },
        { typeof(short?), static value => ((short?)value).Value },
        { typeof(sbyte?), static value => ((sbyte?)value).Value },
        { typeof(ulong?), static value => ((ulong?)value).Value },
        { typeof(uint?), static value => ((uint?)value).Value },
        { typeof(ushort?), static value => ((ushort?)value).Value },
        { typeof(byte?), static value => ((byte?)value).Value }
    };

        internal decimal Number { get; } = number;

        public override string FormatErrorMessage(string name)
        {
            return string.Format(CultureInfo.CurrentCulture, ErrorMessageString, name, Number);
        }

        public override bool IsValid(object value)
        {
            if (value is null)
            {
                return true;
            }

            var valueType = value.GetType();
            if (_defaultDecimalConverters.TryGetValue(valueType, out var converter))
            {
                var convertedValue = converter.Invoke(value);
                return IsValidValue(convertedValue, comparison);
            }

            if (valueType == typeof(string))
            {
                return decimal.TryParse((string)value, out var parsedValue)
                    && IsValidValue(parsedValue, comparison);
            }

            return false;
        }

        private bool IsValidValue(decimal val, int comparison) => comparison switch
        {
            -1 => val <= Number,
            1 => val >= Number,
            _ => val == Number
        };
    }

    internal sealed class NumberMaxAttribute(decimal max) : NumberCompareAttribute(max, -1) { }

    internal sealed class OldNumberMaxAttribute(decimal max) : ValidationAttribute
    {
        internal decimal Max { get; } = max;

        public override string FormatErrorMessage(string name)
        {
            return string.Format(CultureInfo.CurrentCulture, ErrorMessageString, name, Max);
        }

        public override bool IsValid(object value)
        {
            if (value == null)
            {
                return true;
            }

            var type = value.GetType();
            var typeMinValue = GetMinValueString(type);

            var attribute = new RangeAttribute(type, typeMinValue, Max.ToString());

            return attribute.IsValid(value);
        }

        private static string GetMinValueString(Type type)
        {
            if (type == typeof(byte)) return byte.MinValue.ToString();
            if (type == typeof(short)) return short.MinValue.ToString();
            if (type == typeof(int)) return int.MinValue.ToString();
            if (type == typeof(long)) return long.MinValue.ToString();
            if (type == typeof(float)) return float.MinValue.ToString();
            if (type == typeof(double)) return double.MinValue.ToString();
            if (type == typeof(sbyte)) return sbyte.MinValue.ToString();
            if (type == typeof(ushort)) return ushort.MinValue.ToString();
            if (type == typeof(uint)) return uint.MinValue.ToString();
            if (type == typeof(ulong)) return ulong.MinValue.ToString();
            if (type == typeof(decimal)) return decimal.MinValue.ToString();

            return null;
        }
    }
}
```

### 📝 Changelog

<!--
Describe changes from userside, and list all potential break changes or other risks.
--->

| Language   | Changelog |
| ---------- | --------- |
| 🇺🇸 English |     rewrite internal number validation attributes     |
| 🇨🇳 Chinese |    重写内部数字相关的验证特性       |

### ☑️ Self Check before Merge

⚠️ Please check all items below before review. ⚠️

- [x] Doc is updated/provided or not needed
- [x] Demo is updated/provided or not needed
- [x] Changelog is provided or not needed